+++ jmespath/__init__.py
@@ -1,5 +1,7 @@
 from jmespath import parser
 from jmespath.visitor import Options
+from jmespath.budget import BudgetLimits
+from jmespath.budget import current_context
 
 __version__ = '1.1.0'